+++ src/parsing.rs
@@ -21,7 +21,7 @@ fn parse_detail_url(vocab: &str, reading: &str) -> impl Parser<char, String, Err
         just('"')
             .ignore_then(just("/vocabulary/"))
             .then(digits(10))
-            .then(just(format!("/{vocab}/{reading}")))
+            .then(just(format!("/{vocab}/{reading}")).or(just(format!("/{vocab}"))))
             .map(|((a, b), c)| format!("{a}{b}{c}")),
     )
     .map(|(_a, b)| b)
@@ -62,6 +62,15 @@ mod tests {
         let example = r#"href or whatever idc "/vocabulary/1259620/見事/みごと?lang=english#a""#;
         let parsed = parse_detail_url("見事", "みごと").parse(example).unwrap();
         assert_eq!("/vocabulary/1259620/見事/みごと", parsed)
+    }
+
+    #[test]
+    fn parse_detail_url_kana_only_test() {
+        let example = r#"href="/vocabulary/1008290/てっきり?lang=english""#;
+        let parsed = parse_detail_url("てっきり", "てっきり")
+            .parse(example)
+            .unwrap();
+        assert_eq!("/vocabulary/1008290/てっきり", parsed)
     }
 
     #[test]