+++ Changelog.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.0] - 2022-09-05
 ### Added
 - auto_forget option: automatically mark added card as unknown
 - open a card in browser by clicking "View added note (alt-v)" after adding it in yomichan
@@ -13,7 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - improve error message when duplicate detection is on
 - auto features not working with kana only cards
-
 
 ## [0.4.1] - 2022-08-21
 ### Fixed